+++ packages/verifiable-credentials/test/create-credentials.test.ts
@@ -1,5 +1,5 @@
 const assert = require('assert');
-
+require('dotenv').config();
 import Credentials from '../src/credentials';
 import config from './config';
 import { connectAccount } from './utils';
@@ -21,14 +21,15 @@ describe('Credential tests', function () {
         });
 
         it('Verify Credential JWT was created correctly', async function () {
-            const credential: any = await credentialSdk.createCredentialJWT({
+            const didJwtVc: any = await credentialSdk.createCredentialJWT({
                 subjectId: config.SUBJECT_DID,
-                data: config.CREDENTIAL_DATA_PAYLOAD,
+                schema: config.SCHEMA_SBT,
+                data: config.CREDENTIAL_DATA,
                 context: appContext
             });
 
             // Decode the credentialSdk
-            const decodedCredential = await credentialSdk.verifyCredential(credential.didJwtVc, {
+            const decodedCredential = await credentialSdk.verifyCredential(didJwtVc, {
                 rpcUrl: config.DID_CLIENT_CONFIG.rpcUrl
             })
 
@@ -41,47 +42,22 @@ describe('Credential tests', function () {
             assert.equal(payload.iss, issuer.did, 'Credential issuer matches expected DID')
 
             // Verify the data matches the schema
-            const record = vc.credentialSubject
-            const schema = await appContext.getClient().getSchema(record.schema)
-            const isValid = await schema.validate(config.CREDENTIAL_DATA_PAYLOAD);
+            const schema = await appContext.getClient().getSchema(vc.credentialSchema.id)
+            const isValid = await schema.validate(config.CREDENTIAL_DATA)
             assert.equal(true, isValid, 'Credential subject successfully validates against the schema');
 
-            // Verify the "Verifiable Credential"
-            delete config.CREDENTIAL_DATA_PAYLOAD['didJwtVc']
-
-            assert.deepEqual(vc.credentialSubject, config.CREDENTIAL_DATA_PAYLOAD, 'Credential data is valid');
+            assert.deepEqual(vc.credentialSubject, config.CREDENTIAL_DATA, 'Credential data is valid');
             assert.deepEqual(issuer.did, vc.issuer, 'Issuer matches expected DID');
-            assert.equal(vc.credentialSchema.id, config.CREDENTIAL_DATA_PAYLOAD.schema, 'Credential schema is correct')
+            assert.equal(vc.credentialSchema.id, config.SCHEMA_SBT, 'Credential schema is correct')
             assert.equal(vc.sub, config.SUBJECT_DID, 'Credential subject matches expected DID')
-        });
-        it('Check the schema is a credential schema type', async function () {
-            const credential: any = await credentialSdk.createCredentialJWT({
-                subjectId: config.SUBJECT_DID,
-                data: config.CREDENTIAL_DATA_PAYLOAD,
-                context: appContext
-            });
-
-            // Decode the credentialSdk
-            const decodedCredential = await credentialSdk.verifyCredential(credential.didJwtVc, {
-                rpcUrl: config.DID_CLIENT_CONFIG.rpcUrl
-            })
-
-            // Obtain the payload, that contains the verifiable credentialSdk (.vc)
-            const payload = decodedCredential.payload
-            const vc = payload.vc
-
-            // Verify the data matches the schema
-            const record = vc.credentialSubject
-            const schema = await appContext.getClient().getSchema(record.schema)
-            const schemaJson = await schema.getSpecification();
-            assert.ok(schemaJson.properties.didJwtVc, 'schemaJson has the didJwtVc attribute');
         });
         it('Unable to create credential with invalid schema data', async function () {
             const promise = new Promise((resolve, rejects) => {
                 credentialSdk.createCredentialJWT({
                     subjectId: config.SUBJECT_DID,
-                    data: config.INVALID_CREDENTIAL_DATA,
-                    context: appContext,
+                schema: config.SCHEMA_SBT,
+                data: config.INVALID_CREDENTIAL_DATA,
+                context: appContext
                 }).then(rejects, resolve)
             })
             const result = await promise
@@ -92,8 +68,9 @@ describe('Credential tests', function () {
             const promise = new Promise((resolve, rejects) => {
                 credentialSdk.createCredentialJWT({
                     subjectId: config.SUBJECT_DID,
+                    schema: '',
                     data: {},
-                    context: appContext,
+                    context: appContext
                 }).then(rejects, resolve)
             })
             const result = await promise
@@ -103,14 +80,15 @@ describe('Credential tests', function () {
         it('Ensure expired expiration date is respected', async () => {
             // Set an expiry date to the past
             const expirationDate = '2000-02-14T04:27:05.467Z'
-            const credential: any = await credentialSdk.createCredentialJWT({
+            const didJwtVc: any = await credentialSdk.createCredentialJWT({
                 subjectId: config.SUBJECT_DID,
+                schema: config.SCHEMA_SBT,
                 data: config.CREDENTIAL_DATA,
                 context: appContext,
                 options: { expirationDate }
             });
 
-            const decodedCredential = await credentialSdk.verifyCredential(credential.didJwtVc, {
+            const decodedCredential = await credentialSdk.verifyCredential(didJwtVc, {
                 rpcUrl: config.DID_CLIENT_CONFIG.rpcUrl
             })
             assert.equal(decodedCredential, false, 'Credential is not valid')
@@ -119,15 +97,15 @@ describe('Credential tests', function () {
         it('Ensure issuanceDate generated in VC is same in the date options', async () => {
             const issuanceDate = '2022-02-14T04:27:05.467Z';
 
-            const credential: any = await credentialSdk.createCredentialJWT({
+            const didJwtVc: any = await credentialSdk.createCredentialJWT({
                 subjectId: config.SUBJECT_DID,
+                schema: config.SCHEMA_SBT,
                 data: config.CREDENTIAL_DATA,
                 context: appContext,
                 options: { issuanceDate }
             });
 
-            const decodedCredential = await credentialSdk.verifyCredential(credential.didJwtVc, {
-                name: <string> EnvironmentType.TESTNET,
+            const decodedCredential = await credentialSdk.verifyCredential(didJwtVc, {
                 rpcUrl: config.DID_CLIENT_CONFIG.rpcUrl
             })
 
@@ -141,13 +119,14 @@ describe('Credential tests', function () {
             assert.deepEqual(formattedIssuanceVCDate, formattedIssuanceDate, 'issuanceDate options matches generated VC date ');
         });
         it('Ensure issuanceDate generated in VC is within 10secs', async () => {
-            const credential: any = await credentialSdk.createCredentialJWT({
+            const didJwtVc: any = await credentialSdk.createCredentialJWT({
                 subjectId: config.SUBJECT_DID,
+                schema: config.SCHEMA_SBT,
                 data: config.CREDENTIAL_DATA,
-                context: appContext,
+                context: appContext
             });
 
-            const decodedCredential = await credentialSdk.verifyCredential(credential.didJwtVc, {
+            const decodedCredential = await credentialSdk.verifyCredential(didJwtVc, {
                 rpcUrl: config.DID_CLIENT_CONFIG.rpcUrl
             })
             const payload = decodedCredential.payload
@@ -165,14 +144,15 @@ describe('Credential tests', function () {
             const expirationDate = '2000-02-14T04:27:05.467Z';
             const issuanceDate = '2022-02-14T04:27:05.467Z';
             const currentDateTime = '2024-02-14T04:27:05.467Z';
-            const credential: any = await credentialSdk.createCredentialJWT({
+            const didJwtVc: any = await credentialSdk.createCredentialJWT({
                 subjectId: config.SUBJECT_DID,
+                schema: config.SCHEMA_SBT,
                 data: config.CREDENTIAL_DATA,
                 context: appContext,
                 options: { issuanceDate, expirationDate }
             });
 
-            await credentialSdk.verifyCredential(credential.didJwtVc, {
+            await credentialSdk.verifyCredential(didJwtVc, {
                 rpcUrl: config.DID_CLIENT_CONFIG.rpcUrl
             }, currentDateTime);
 
@@ -181,8 +161,9 @@ describe('Credential tests', function () {
         it('Ensure valid expiration date is respected', async () => {
             // Set an expiry date to the future
             const expirationDate = '2060-02-14T04:27:05.467Z'
-            const credential: any = await credentialSdk.createCredentialJWT({
+            const didJwtVc: any = await credentialSdk.createCredentialJWT({
                 subjectId: config.SUBJECT_DID,
+                schema: config.SCHEMA_SBT,
                 data: config.CREDENTIAL_DATA,
                 context: appContext,
                 options: {
@@ -190,7 +171,7 @@ describe('Credential tests', function () {
                 }
             });
 
-            const decodedCredential = await credentialSdk.verifyCredential(credential.didJwtVc, {
+            const decodedCredential = await credentialSdk.verifyCredential(didJwtVc, {
                 rpcUrl: config.DID_CLIENT_CONFIG.rpcUrl
             })
             assert.ok(decodedCredential, 'Credential is valid')