+++ packages/helpers/CHANGELOG.md
@@ -1,5 +1,4 @@
-
-2023-03-07 (v2.1.4)
+2023-03-10 (v2.2.0)
 -------------------
 
 - Support `wrapUri()` method
@@ -7,6 +6,8 @@
 - Add explode DID helper method
 - Support deep attributes in Verida URIs
 - Add unit tests
+- Support `deepAttributes` in `buildVeridaUri()`
+- Unwrap `deepAttributes` in `fetchVeridaUri()` before attempting to decrypt
 
 2023-02-16 (v2.1.1)
 -------------------