+++ packages/vda-sbt-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+2023-03-13 (v2.1.5)
+-------------------
+- Updated `isSBTClaimed()` function
+- Added `getClaimedSBTList()` function
+- Updated test code
+
 2023-03-07 (v2.1.4)
 -------------------